// ui/src/client/types.gen.ts
export type ItemKind = 'node' | 'edge' | 'workflow';
// ...
export type WorkflowError = {
    kind: ItemKind;
    /**
     * Id
     */
    id: string | null;
    /**
     * Field
     */
    field: string | null;
    /**
     * Message
     */
    message: string;
};